Fixed some styles and better folder structure
