+++ src/Network/index.ts
@@ -1,0 +1,11 @@
+import { API } from './api.ts';
+import HttpClient from './http-client/http-client.ts';
+import {
+  PostReqOptions,
+  SearchResponse,
+  IHttpClient,
+} from './http-client/http-client.types.ts';
+
+export { API };
+export { HttpClient };
+export type { PostReqOptions, SearchResponse, IHttpClient };
