--- src/Network/network.txt
@@ -1,1 +1,0 @@
-This is a mock file, please remove it when you have a network library to add to the folder
